+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # CHANGELOG
+## 0.0.46
+- [scoreboard-spigot] Add support for minecraft 1.21.6
+- [common] Bump Minecraft versions from 1.21.6 to 1.21.10
+
+## 0.0.45
+- [common] Add DNSLookup
+
 ## 0.0.44
 - [commands] Add ArgumentConverter